[fix] remove master, fix typos
