+++ mmocr/models/textrecog/plugins/__init__.py
@@ -1,4 +1,4 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from .common_plugins import Maxpool2d
+from .common import Maxpool2d
 
 __all__ = ['Maxpool2d']
